+++ zingo_common_components/Cargo.toml
@@ -4,10 +4,9 @@ version = "0.1.0"
 edition = "2024"
 
 [features]
-librustzcash_zebra_compatibility = [ "zebra-chain" ]
 for_test = []
 
 [dependencies]
 thiserror.workspace = true
 zcash_protocol.workspace = true
-zebra-chain = { workspace = true, optional = true }
+zebra-chain = { workspace = true }
